+++ tests/run.rb
@@ -4,8 +4,8 @@ require_relative 'class_tests.rb'
 @tests = Tests.new
 
 # Add tests here
-require_relative 'test_full_yaml.rb'
-require_relative 'test_required_yaml.rb'
+require_relative 'test_full.rb'
+require_relative 'test_required.rb'
 require_relative 'test_utilities.rb'
 
 # Run tests
