+++ tests/test_full.rb
@@ -1,5 +1,5 @@
 full_config_file = 'tests/test_full.yaml'
-full_dns_hosts_file = File.join(File.dirname(__FILE__), 'test_full_yaml_dns_hosts.txt')
+full_dns_hosts_file = File.join(File.dirname(__FILE__), 'test_full_dns_hosts.txt')
 testing = Config.new(full_config_file, full_dns_hosts_file)
 
 @tests.assertions.push(*[
